Refactor: MVC패턴 적용
