+++ src/main/java/com/example/Level6/Model/Cart.java
@@ -1,4 +1,4 @@
-package com.example.Level6;
+package com.example.Level6.Model;
 
 import java.util.Collections;
 import java.util.List;
@@ -9,7 +9,7 @@ public class Cart {
     private final List<MenuItem> orderList;
 
     //생성자
-    public Cart(List<MenuItem> orderList){
+    public Cart(List<MenuItem> orderList) {
         this.orderList = orderList;
     }
 
@@ -19,16 +19,16 @@ public class Cart {
     }
 
     //주문 상품 추가
-    public void addItem (MenuItem item){
+    public void addItem(MenuItem item) {
         orderList.add(item);
     }
 
     //카트
-    public void clearItem(){
+    public void clearItem() {
         orderList.clear();
     }
 
-    public boolean checkItem (){
+    public boolean checkItem() {
         return !orderList.isEmpty(); //장바구니가 비었는지 확인
     }
